Graphics are back, but skeletons are disabled. Comments have also been fixed across all files.

+++ src/settingsProgram.h
@@ -2,6 +2,10 @@
 #define settingsProgram_h
 
 
+// Included for "size_t".
+#include <stddef.h>
+
+
 #define ASPECT_RATIO_X 4
 #define ASPECT_RATIO_Y 3
 
@@ -10,8 +14,8 @@
 #define UPDATE_RATE 60.f
 #define FRAME_RATE 120.f
 
-//We'll probably get undefined results if this is less than or equal to 0.
-#define NUM_LOOKBACK_STATES ((unsigned int)UPDATE_RATE)
+// We'll probably get undefined results if this is less than or equal to 0.
+#define NUM_LOOKBACK_STATES ((size_t)UPDATE_RATE)
